requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
duckdb==0.7.0
streamlit==1.18.0
minio==7.1.13
pandas==1.5.3
airflow-provider-duckdb==0.0.2
astro-sdk-python==1.5.0
spacy==3.7.1
textblob==0.17.1
tensorflow==2.14.0
google-api-python-client==2.103.0
transformers==4.34.0
torch==2.1.0 
torchvision==0.16.0
altair==4.2.2